requirements.txt
python-docx==1.0.0
